Converting int to byte to save some space

+++ src/fighters.h
@@ -1,7 +1,7 @@
 #ifndef FIGHTERS_H
 #define FIGHTERS_H
 
-int fighter_index_first = -1;
-int fighter_index_second = -1;
+byte fighter_index_first = -1;
+byte fighter_index_second = -1;
 
 #endif // FIGHTERS_H
